refactor: use PascalCase for component files
