+++ eslint.config.cjs
@@ -12,10 +12,9 @@ module.exports = [
       '.DS_Store',
       'CLAUDE.md',
       'src/templates/**',
-      'src/components/practice/routines/template.vue',
-      'src/components/random/template.vue',
-      'src/application.vue',
-      'src/components/navigation/template.vue',
+      'src/components/practice/routines/Routines.vue',
+      'src/components/random/Random.vue',
+      'src/App.vue',
       'eslint.config.cjs',
       'eslint-wide-formatter.js',
       'vite.config.js',
