+++ package.json
@@ -12,7 +12,7 @@
     "preview": "vite preview --port 4173",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs -f ./eslint-wide-formatter.js",
     "fix": "npm run lint -- --fix",
-    "test": "node --test --experimental-test-coverage && npm run lint",
+    "test": "node --test --experimental-test-coverage tests/**/*.js && npm run lint",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d dist"
   },
