+++ src/components/adventure/destinations/README.md
@@ -3,4 +3,4 @@
 Page listing travel destinations from `destinations.json`.
 
 ## Component
-- `template.vue` – iterates over destinations list. No props.
+- `Destinations.vue` – iterates over destinations list. No props.
