+++ src/components/adventure/overview/README.md
@@ -3,4 +3,4 @@
 Introductory article for the adventure section.
 
 ## Component
-- `template.vue` – static page content. No props or events.
+- `Overview.vue` – static page content. No props or events.
