+++ src/components/entertainment/anime/Anime.vue
@@ -19,8 +19,8 @@
 -->
 
 <script setup>
-import ArticleTemplate from '../../../templates/article.vue';
-import ActionsTemplate from '../../../templates/actions.vue';
+import ArticleTemplate from '../../../templates/Article.vue';
+import ActionsTemplate from '../../../templates/Actions.vue';
 import animeList from './data.json';
 
 defineOptions({
