+++ src/components/entertainment/anime/README.md
@@ -3,5 +3,5 @@
 Checklist of anime titles.
 
 ## Files
-- `template.vue` – renders suggestions from `data.json`.
+- `Anime.vue` – renders suggestions from `data.json`.
 - `data.json` – list of anime titles and completion state.
