+++ src/components/entertainment/movies/README.md
@@ -3,4 +3,4 @@
 Placeholder page for movie suggestions.
 
 ## Component
-- `template.vue` – static content. No props or events.
+- `Movies.vue` – static content. No props or events.
