+++ src/components/entertainment/overview/README.md
@@ -3,4 +3,4 @@
 Landing page for entertainment content.
 
 ## Component
-- `template.vue` – static description of the entertainment section.
+- `Overview.vue` – static description of the entertainment section.
