+++ src/components/experiments/README.md
@@ -3,5 +3,5 @@
 Personal experiments and notes.
 
 ## Component
-- `template.vue` – displays experiment write-ups and images from the `assets/` folder.
+- `Experiments.vue` – displays experiment write-ups and images from the `assets/` folder.
 - `suggestions.json` – optional checklist for future experiments.
