+++ src/components/footer/README.md
@@ -3,4 +3,4 @@
 Simple page footer with attribution.
 
 ## Component
-- `template.vue` – displays copyright and author link. No props or events.
+- `AppFooter.vue` – displays copyright and author link. No props or events.
