+++ src/components/header/README.md
@@ -3,4 +3,4 @@
 Static hero section displayed at the top of every page.
 
 ## Component
-- `template.vue` – contains the project tagline and introductory text. No props or events.
+- `AppHeader.vue` – contains the project tagline and introductory text. No props or events.
