+++ src/components/home/README.md
@@ -3,4 +3,4 @@
 Landing page with introductory articles and an actions widget.
 
 ## Component
-- `template.vue` – renders static articles and an `ActionsTemplate` bound to a predefined list ID.
+- `Home.vue` – renders static articles and an `ActionsTemplate` bound to a predefined list ID.
