+++ src/components/ikigai/README.md
@@ -3,4 +3,4 @@
 Self-reflection notes and action list.
 
 ## Component
-- `template.vue` – combines `ActionsTemplate` with a suggestions checklist loaded from `suggestions.json`.
+- `Ikigai.vue` – combines `ActionsTemplate` with a suggestions checklist loaded from `suggestions.json`.
