+++ src/components/ippo/README.md
@@ -3,4 +3,4 @@
 Work-in-progress page for personal projects.
 
 ## Component
-- `template.vue` – placeholder article content. No props or events.
+- `Ippo.vue` – placeholder article content. No props or events.
