+++ src/components/literature/books/README.md
@@ -3,5 +3,5 @@
 Track reading progress using action lists.
 
 ## Files
-- `template.vue` – displays book-related actions.
+- `Books.vue` – displays book-related actions.
 - `actions.sql` – SQL schema for storing book actions in Supabase.
