+++ src/components/literature/overview/README.md
@@ -3,4 +3,4 @@
 Landing page for literature-related content.
 
 ## Component
-- `template.vue` – static introduction. No props or events.
+- `Overview.vue` – static introduction. No props or events.
