+++ src/components/navigation/README.md
@@ -3,7 +3,7 @@
 Responsive site navigation bar.
 
 ## Component
-- `template.vue` – builds menus from `useAuthentication` helpers and emits `showAuthentication` when login is requested.
+- `AppNavigation.vue` – builds menus from `useAuthentication` helpers and emits `showAuthentication` when login is requested.
 
 ## Props & Events
 No props. Emits `showAuthentication` to trigger the password modal.
