+++ src/components/nutrition/ingredients/README.md
@@ -3,5 +3,5 @@
 Interactive ingredient explorer.
 
 ## Files
-- `template.vue` – displays a searchable list of ingredients.
+- `Ingredients.vue` – displays a searchable list of ingredients.
 - `ingredients.json` – ingredient data with nutritional info.
