+++ src/components/practice/routines/README.md
@@ -3,6 +3,6 @@
 Routines and habit trackers.
 
 ## Files
-- `template.vue` – renders routines and activities.
+- `Routines.vue` – renders routines and activities.
 - `habits.json`, `activities.json` – data sources for routines.
 - `schemas/` – database schema references.
