+++ src/components/random/README.md
@@ -3,4 +3,4 @@
 Miscellaneous notes and pages.
 
 ## Component
-- `template.vue` – placeholder page with arbitrary content. No props or events.
+- `Random.vue` – placeholder page with arbitrary content. No props or events.
