+++ src/templates/Suggestions.vue
